Start to remove useless code for CssAnalyzerBridgeServer
