+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/CssAnalyzerBridgeServer.java
@@ -17,43 +17,33 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.nodejs.server;
+package org.sonar.css.plugin.server;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.Duration;
-import java.util.Optional;
-import java.util.stream.Stream;
+import javax.annotation.Nullable;
 import okhttp3.HttpUrl;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
-import org.sonar.api.batch.fs.FileSystem;
 import org.sonar.api.batch.sensor.SensorContext;
-import org.sonar.api.config.Configuration;
 import org.sonar.api.utils.log.Logger;
 import org.sonar.api.utils.log.Loggers;
 import org.sonar.api.utils.log.Profiler;
+import org.sonar.css.plugin.server.exception.ServerAlreadyFailedException;
 import org.sonarsource.nodejs.NodeCommand;
 import org.sonarsource.nodejs.NodeCommandBuilder;
 import org.sonarsource.nodejs.NodeCommandException;
 
-import static java.util.Collections.emptyList;
+public class CssAnalyzerBridgeServer implements AnalyzerBridgeServer {
 
-public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
-
-  private static final Logger LOG = Loggers.get(AnalyzerBridgeServerImpl.class);
+  private static final Logger LOG = Loggers.get(CssAnalyzerBridgeServer.class);
   private static final Profiler PROFILER = Profiler.createIfDebug(LOG);
-
-  // SonarLint should pass in this property an absolute path to the directory containing TypeScript dependency
-  private static final String TYPESCRIPT_DEPENDENCY_LOCATION_PROPERTY = "sonar.typescript.internal.typescriptLocation";
 
   private static final int DEFAULT_TIMEOUT_SECONDS = 60;
   // internal property to set "--max-old-space-size" for Node process running this server
@@ -67,17 +57,15 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
   private int port;
   private NodeCommand nodeCommand;
   private boolean failedToStart;
-  private Configuration configuration;
 
   // Used by pico container for dependency injection
   @SuppressWarnings("unused")
-  public AnalyzerBridgeServerImpl(Configuration configuration, NodeCommandBuilder nodeCommandBuilder, Bundle bundle) {
-    this(configuration, nodeCommandBuilder, DEFAULT_TIMEOUT_SECONDS, bundle);
+  public CssAnalyzerBridgeServer(NodeCommandBuilder nodeCommandBuilder, Bundle bundle) {
+    this(nodeCommandBuilder, DEFAULT_TIMEOUT_SECONDS, bundle);
   }
 
-  protected AnalyzerBridgeServerImpl(Configuration configuration, NodeCommandBuilder nodeCommandBuilder, int timeoutSeconds,
+  protected CssAnalyzerBridgeServer(NodeCommandBuilder nodeCommandBuilder, int timeoutSeconds,
                          Bundle bundle) {
-    this.configuration = configuration;
     this.nodeCommandBuilder = nodeCommandBuilder;
     this.timeoutSeconds = timeoutSeconds;
     this.bundle = bundle;
@@ -132,21 +120,7 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
       .getInt(MAX_OLD_SPACE_SIZE_PROPERTY)
       .ifPresent(nodeCommandBuilder::maxOldSpaceSize);
 
-    if (shouldDetectTypeScript(context.fileSystem())) {
-      Optional<Path> typeScriptLocation = getTypeScriptLocation(context.fileSystem().baseDir());
-      if (typeScriptLocation.isPresent()) {
-        LOG.info("Using TypeScript at: '{}'", typeScriptLocation.get());
-        nodeCommandBuilder.addToNodePath(typeScriptLocation.get().toAbsolutePath());
-      } else {
-        LOG.info("TypeScript dependency was not found inside project directory, Node.js will search TypeScript using " +
-          "module resolution algorithm; analysis will fail without TypeScript.");
-      }
-    }
     nodeCommand = nodeCommandBuilder.build();
-  }
-
-  private static boolean shouldDetectTypeScript(FileSystem fileSystem) {
-    return fileSystem.hasFiles(TypeScriptSensor.filePredicate(fileSystem));
   }
 
   @Override
@@ -169,10 +143,14 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
     }
   }
 
+  public AnalysisResponse analyze(AnalysisRequest request) throws IOException {
+    return analyze("analyze-css", request, AnalysisResponse.class);
+  }
+
   @Override
-  public AnalysisResponse analyze(String endpoint, AnalysisRequest request) throws IOException {
+  public <REQ, RES> RES analyze(String endpoint, REQ request, Class<RES> cls) throws IOException {
     String json = GSON.toJson(request);
-    return response(request(json, endpoint), request.filePath);
+    return response(request(json, endpoint), cls);
   }
 
   private String request(String json, String endpoint) throws IOException {
@@ -187,11 +165,11 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
     }
   }
 
-  private static AnalysisResponse response(String result, String filePath) {
+  private static <RES> RES response(String result, Class<RES> cls) {
     try {
-      return GSON.fromJson(result, AnalysisResponse.class);
+      return GSON.fromJson(result, cls);
     } catch (JsonSyntaxException e) {
-      String msg = "Failed to parse response for file " + filePath + ": \n-----\n" + result + "\n-----\n";
+      String msg = "Failed to parse response for file " + /*TODO*/ ": \n-----\n" + result + "\n-----\n";
       LOG.error(msg, e);
       throw new IllegalStateException("Failed to parse response", e);
     }
@@ -218,56 +196,6 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
   }
 
   @Override
-  public boolean newTsConfig() {
-    Request request = new Request.Builder()
-      .url(url("new-tsconfig"))
-      .post(RequestBody.create(null, ""))
-      .build();
-    try (Response response = client.newCall(request).execute()) {
-      String body = response.body().string();
-      return "OK!".equals(body);
-    } catch (IOException e) {
-      LOG.error("Failed to post new-tsconfig", e);
-    }
-    return false;
-  }
-
-  public TsConfigResponse tsConfigFiles(String tsconfigAbsolutePath) {
-    String result = null;
-    try {
-      TsConfigRequest tsConfigRequest = new TsConfigRequest(tsconfigAbsolutePath);
-      result = request(GSON.toJson(tsConfigRequest), "tsconfig-files");
-      return GSON.fromJson(result, TsConfigResponse.class);
-    } catch (IOException e) {
-      LOG.error("Failed to request files for tsconfig: " + tsconfigAbsolutePath, e);
-    } catch (JsonSyntaxException e) {
-      LOG.error("Failed to parse response when requesting files for tsconfig: " + tsconfigAbsolutePath + ": \n-----\n" + result + "\n-----\n");
-    }
-    return new TsConfigResponse(emptyList(), result, null);
-  }
-
-  @Override
-  public TsConfigFile loadTsConfig(String filename) {
-    AnalyzerBridgeServer.TsConfigResponse tsConfigResponse = tsConfigFiles(filename);
-    if (tsConfigResponse.error != null) {
-      LOG.error(tsConfigResponse.error);
-      if (tsConfigResponse.errorCode == AnalyzerBridgeServer.ParsingErrorCode.MISSING_TYPESCRIPT) {
-        TypeScriptSensor.logMissingTypescript();
-        throw new MissingTypeScriptException();
-      }
-    }
-    return new TsConfigFile(filename, tsConfigResponse.files == null ? emptyList() : tsConfigResponse.files);
-  }
-
-  @Override
-  public void clean() {
-    if (nodeCommand != null) {
-      nodeCommand.destroy();
-      nodeCommand = null;
-    }
-  }
-
-  @Override
   public String getCommandInfo() {
     if (nodeCommand == null) {
       return "Node.js command to start " + bundle.name() + " server was not built yet.";
@@ -286,6 +214,13 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
     clean();
   }
 
+  void clean() {
+    if (nodeCommand != null) {
+      nodeCommand.destroy();
+      nodeCommand = null;
+    }
+  }
+
   private HttpUrl url(String endpoint) {
     HttpUrl.Builder builder = new HttpUrl.Builder();
     return builder
@@ -296,28 +231,19 @@ public class AnalyzerBridgeServerImpl implements AnalyzerBridgeServer {
       .build();
   }
 
-  private Optional<Path> getTypeScriptLocation(File baseDir) throws IOException {
-    // we have to use global Configuration and not SensorContext#config to lookup typescript set from vscode extension
-    // see https://jira.sonarsource.com/browse/SLCORE-250
-    Optional<String> typeScriptLocationProperty = configuration.get(TYPESCRIPT_DEPENDENCY_LOCATION_PROPERTY);
-    if (typeScriptLocationProperty.isPresent()) {
-      LOG.debug("TypeScript location set via property {}={}", TYPESCRIPT_DEPENDENCY_LOCATION_PROPERTY, typeScriptLocationProperty.get());
-      return Optional.of(Paths.get(typeScriptLocationProperty.get()));
-    }
-    LOG.debug("Looking for TypeScript recursively in {}", baseDir.getAbsolutePath());
-    try (Stream<Path> files = Files.walk(baseDir.toPath())) {
-      return files
-        .filter(p -> p.toFile().isDirectory() && p.endsWith("node_modules/typescript"))
-        .findFirst()
-        .map(Path::getParent);
+  public static class AnalysisRequest {
+    String filePath;
+    @Nullable
+    String configFile;
+
+    public AnalysisRequest(String filePath, @Nullable String configFile) {
+      this.filePath = filePath;
+      this.configFile = configFile;
     }
   }
 
-  static class TsConfigRequest {
-    final String tsconfig;
-
-    TsConfigRequest(String tsconfig) {
-      this.tsconfig = tsconfig;
-    }
+  public static class AnalysisResponse {
+    String[] issues;
   }
+
 }
