+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/CssAnalyzerBundle.java
@@ -17,11 +17,22 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.nodejs.server;
+package org.sonar.css.plugin.server;
 
-public class MissingTypeScriptException extends RuntimeException {
+import org.sonar.api.utils.TempFolder;
 
-  public MissingTypeScriptException() {
-    super("Missing TypeScript dependency");
+public class CssAnalyzerBundle extends BundleImpl {
+
+  // this archive is created in eslint-bridge module
+  private static final String BUNDLE_LOCATION = "/css-analyzer-bridge-1.0.0.tgz";
+  private static final String DEPLOY_LOCATION = "css-analyzer-bridge-bundle";
+
+  public CssAnalyzerBundle(TempFolder tempFolder) {
+    this(tempFolder, BUNDLE_LOCATION);
   }
+
+  CssAnalyzerBundle(TempFolder tempFolder, String bundleLocation) {
+    super(tempFolder.newDir(DEPLOY_LOCATION).toPath(), bundleLocation, "css-analyzer-bridge");
+  }
+
 }
