+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/NetUtils.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.nodejs.server;
+package org.sonar.css.plugin.server;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
