+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/package-info.java
@@ -17,16 +17,5 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.nodejs.server;
-
-import java.io.IOException;
-import org.sonarsource.nodejs.BundlePathResolver;
-
-public interface Bundle extends BundlePathResolver {
-
-  void deploy() throws IOException;
-
-  String startServerScript();
-
-  String name();
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package org.sonar.css.plugin.server;
