+++ sonar-css-plugin/src/test/java/org/sonar/css/plugin/server/CssAnalyzerBridgeServerTest.java
@@ -1,0 +1,276 @@
+/*
+ * SonarCSS
+ * Copyright (C) 2018-2019 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.css.plugin.server;
+
+import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+import org.sonar.api.batch.fs.internal.DefaultInputFile;
+import org.sonar.api.batch.fs.internal.TestInputFileBuilder;
+import org.sonar.api.batch.sensor.internal.SensorContextTester;
+import org.sonar.api.config.internal.MapSettings;
+import org.sonar.api.utils.internal.JUnitTempFolder;
+import org.sonar.api.utils.log.LogTester;
+import org.sonar.css.plugin.server.CssAnalyzerBridgeServer.AnalysisRequest;
+import org.sonar.css.plugin.server.exception.ServerAlreadyFailedException;
+import org.sonarsource.nodejs.NodeCommand;
+import org.sonarsource.nodejs.NodeCommandBuilder;
+import org.sonarsource.nodejs.NodeCommandException;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.mock;
+import static org.sonar.api.utils.log.LoggerLevel.DEBUG;
+import static org.sonar.api.utils.log.LoggerLevel.INFO;
+import static org.sonar.api.utils.log.LoggerLevel.WARN;
+
+public class CssAnalyzerBridgeServerTest {
+
+  private static final String START_SERVER_SCRIPT = "startServer.js";
+  private static final int TEST_TIMEOUT_SECONDS = 1;
+
+  @org.junit.Rule
+  public LogTester logTester = new LogTester();
+
+  @org.junit.Rule
+  public final ExpectedException thrown = ExpectedException.none();
+
+  @org.junit.Rule
+  public JUnitTempFolder tempFolder = new JUnitTempFolder();
+
+  private SensorContextTester context;
+  private CssAnalyzerBridgeServer cssAnalyzerBridgeServer;
+  private TestBundle testBundle = new TestBundle(START_SERVER_SCRIPT);
+
+  @Before
+  public void setUp() throws Exception {
+    context = SensorContextTester.create(tempFolder.newDir());
+  }
+
+  @After
+  public void tearDown() throws Exception {
+    cssAnalyzerBridgeServer.clean();
+  }
+
+  @Test
+  public void should_throw_when_not_existing_script() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("NOT_EXISTING.js");
+    cssAnalyzerBridgeServer.deploy();
+
+    thrown.expect(NodeCommandException.class);
+    thrown.expectMessage("Node.js script to start eslint-bridge server doesn't exist:");
+
+    cssAnalyzerBridgeServer.startServer(context);
+  }
+
+  @Test
+  public void should_throw_if_failed_to_build_node_command() throws Exception {
+    NodeCommandBuilder nodeCommandBuilder = mock(NodeCommandBuilder.class, invocation -> {
+      if (NodeCommandBuilder.class.equals(invocation.getMethod().getReturnType())) {
+        return invocation.getMock();
+      } else {
+        throw new NodeCommandException("msg");
+      }
+    });
+
+    cssAnalyzerBridgeServer = new CssAnalyzerBridgeServer(nodeCommandBuilder, TEST_TIMEOUT_SECONDS, testBundle);
+    cssAnalyzerBridgeServer.deploy();
+
+    thrown.expect(NodeCommandException.class);
+    thrown.expectMessage("msg");
+
+    cssAnalyzerBridgeServer.startServer(context);
+  }
+
+  @Test
+  public void should_forward_process_streams() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("logging.js");
+    cssAnalyzerBridgeServer.deploy();
+    cssAnalyzerBridgeServer.startServer(context);
+
+    assertThat(logTester.logs(DEBUG)).contains("testing debug log");
+    assertThat(logTester.logs(WARN)).contains("testing warn log");
+    assertThat(logTester.logs(INFO)).contains("testing info log");
+  }
+
+  @Test
+  public void should_get_answer_from_server() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    cssAnalyzerBridgeServer.deploy();
+    cssAnalyzerBridgeServer.startServer(context);
+
+    DefaultInputFile inputFile = TestInputFileBuilder.create("foo", "foo.js")
+      .setContents("alert('Fly, you fools!')")
+      .build();
+    AnalysisRequest request = new AnalysisRequest(inputFile.absolutePath(), null);
+    assertThat(cssAnalyzerBridgeServer.analyze(request).issues).isEmpty();
+  }
+
+  @Test
+  public void should_get_answer_from_server_for_ts_request() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    cssAnalyzerBridgeServer.deploy();
+    cssAnalyzerBridgeServer.startServer(context);
+
+    DefaultInputFile inputFile = TestInputFileBuilder.create("foo", "foo.ts")
+      .setContents("alert('Fly, you fools!')")
+      .build();
+    DefaultInputFile tsConfig = TestInputFileBuilder.create("foo", "tsconfig.json")
+      .setContents("{\"compilerOptions\": {\"target\": \"es6\", \"allowJs\": true }}")
+      .build();
+    AnalysisRequest request = new AnalysisRequest(inputFile.absolutePath(), null);
+    assertThat(cssAnalyzerBridgeServer.analyze(request).issues).isEmpty();
+  }
+
+  @Test
+  public void should_throw_if_failed_to_start() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("throw.js");
+    cssAnalyzerBridgeServer.deploy();
+
+    thrown.expect(NodeCommandException.class);
+    thrown.expectMessage("Failed to start server (" + TEST_TIMEOUT_SECONDS + "s timeout)");
+
+    cssAnalyzerBridgeServer.startServer(context);
+  }
+
+  @Test
+  public void should_return_command_info() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).isEqualTo("Node.js command to start eslint-bridge server was not built yet.");
+
+    cssAnalyzerBridgeServer.deploy();
+    cssAnalyzerBridgeServer.startServer(context);
+
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).contains("Node.js command to start eslint-bridge was: ", "node", START_SERVER_SCRIPT);
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).doesNotContain("--max-old-space-size");
+  }
+
+  @Test
+  public void should_set_max_old_space_size() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).isEqualTo("Node.js command to start eslint-bridge server was not built yet.");
+
+    cssAnalyzerBridgeServer.deploy();
+    context.setSettings(new MapSettings().setProperty("sonar.javascript.node.maxspace", 2048));
+    cssAnalyzerBridgeServer.startServer(context);
+
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).contains("--max-old-space-size=2048");
+  }
+
+  @Test
+  public void test_isAlive() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    assertThat(cssAnalyzerBridgeServer.isAlive()).isFalse();
+    cssAnalyzerBridgeServer.startServerLazily(context);
+    assertThat(cssAnalyzerBridgeServer.isAlive()).isTrue();
+    cssAnalyzerBridgeServer.clean();
+    assertThat(cssAnalyzerBridgeServer.isAlive()).isFalse();
+  }
+
+  @Test
+  public void test_lazy_start() throws Exception {
+    String alreadyStarted = "eslint-bridge server is up, no need to start.";
+    String starting = "Starting Node.js process to start eslint-bridge server at port";
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("startServer.js");
+    cssAnalyzerBridgeServer.startServerLazily(context);
+    assertThat(logTester.logs(DEBUG).stream().anyMatch(s -> s.startsWith(starting))).isTrue();
+    assertThat(logTester.logs(DEBUG)).doesNotContain(alreadyStarted);
+    logTester.clear();
+    cssAnalyzerBridgeServer.startServerLazily(context);
+    assertThat(logTester.logs(DEBUG).stream().noneMatch(s -> s.startsWith(starting))).isTrue();
+    assertThat(logTester.logs(DEBUG)).contains(alreadyStarted);
+  }
+
+  @Test
+  public void should_throw_special_exception_when_failed_already() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("throw.js");
+    String failedToStartExceptionMessage = "Failed to start server (" + TEST_TIMEOUT_SECONDS + "s timeout)";
+    assertThatThrownBy(() -> cssAnalyzerBridgeServer.startServerLazily(context))
+      .isInstanceOf(NodeCommandException.class)
+      .hasMessage(failedToStartExceptionMessage);
+
+    assertThatThrownBy(() -> cssAnalyzerBridgeServer.startServerLazily(context))
+      .isInstanceOf(ServerAlreadyFailedException.class);
+  }
+
+  @Test
+  public void should_fail_if_bad_json_response() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer("badResponse.js");
+    cssAnalyzerBridgeServer.deploy();
+    cssAnalyzerBridgeServer.startServerLazily(context);
+
+    DefaultInputFile inputFile = TestInputFileBuilder.create("foo", "foo.js")
+      .setContents("alert('Fly, you fools!')")
+      .build();
+    AnalysisRequest request = new AnalysisRequest(inputFile.absolutePath(), null);
+    assertThatThrownBy(() -> cssAnalyzerBridgeServer.analyze(request)).isInstanceOf(IllegalStateException.class);
+    assertThat(context.allIssues()).isEmpty();
+  }
+
+  @Test
+  public void should_not_search_typescript_when_no_ts_file() throws Exception {
+    cssAnalyzerBridgeServer = createCssAnalyzerBridgeServer(START_SERVER_SCRIPT);
+    cssAnalyzerBridgeServer.deploy();
+    Path baseDir = tempFolder.newDir().toPath();
+    SensorContextTester ctx = SensorContextTester.create(baseDir);
+    Path tsDir = baseDir.resolve("dir/node_modules/typescript");
+    Files.createDirectories(tsDir);
+    cssAnalyzerBridgeServer.startServer(ctx);
+    assertThat(cssAnalyzerBridgeServer.getCommandInfo()).doesNotContain("NODE_PATH");
+  }
+
+  private CssAnalyzerBridgeServer createCssAnalyzerBridgeServer(String startServerScript) {
+    return new CssAnalyzerBridgeServer(NodeCommand.builder(), TEST_TIMEOUT_SECONDS, new TestBundle(startServerScript));
+  }
+
+  static class TestBundle implements Bundle {
+
+    final String startServerScript;
+
+    TestBundle(String startServerScript) {
+      this.startServerScript = startServerScript;
+    }
+
+    @Override
+    public String name() {
+      return "eslint-bridge";
+    }
+
+    @Override
+    public void deploy() {
+      // no-op for unit test
+    }
+
+    @Override
+    public String startServerScript() {
+      return "src/test/resources/mock-css-analyzer-bridge/" + startServerScript;
+    }
+
+    @Override
+    public String resolve(String relativePath) {
+      File file = new File("src/test/resources");
+      return new File(file.getAbsoluteFile(), relativePath).getAbsolutePath();
+    }
+  }
+}
