+++ sonar-css-plugin/src/test/resources/mock-css-analyzer-bridge/badResponse.js
