+++ sonar-css-plugin/src/test/resources/mock-css-analyzer-bridge/tsConfigError.js
